+++ packages/xchain-thorchain/src/client.ts
@@ -16,12 +16,12 @@ import { Asset, baseAmount } from '@xchainjs/xchain-util'
 import * as xchainCrypto from '@xchainjs/xchain-crypto'
 
 import { PrivKey, codec, Msg, AccAddress } from 'cosmos-client'
-import { StdTxSignature } from 'cosmos-client/api'
+import { StdTxFee } from 'cosmos-client/api'
 import { StdTx } from 'cosmos-client/x/auth'
 import { MsgSend, MsgMultiSend } from 'cosmos-client/x/bank'
 
 import { AssetRune, MsgNativeTx, DepositParam } from './types'
-import { getDenom, getAsset, getTxsFromHistory, DECIMAL } from './util'
+import { getDenom, getAsset, getTxsFromHistory, DECIMAL, getDenomWithChain, isBroadcastSuccess } from './util'
 
 /**
  * Interface for custom Thorchain client
@@ -29,7 +29,7 @@ import { getDenom, getAsset, getTxsFromHistory, DECIMAL } from './util'
 export interface ThorchainClient {
   validateAddress(address: string): boolean
 
-  deposit({ amount, memo }: DepositParam): Promise<TxHash>
+  deposit(params: DepositParam): Promise<TxHash>
 }
 
 class Client implements ThorchainClient, XChainClient {
@@ -53,13 +53,13 @@ class Client implements ThorchainClient, XChainClient {
     if (phrase) this.setPhrase(phrase)
   }
 
-  purgeClient(): void {
+  purgeClient = (): void => {
     this.phrase = ''
     this.address = ''
     this.privateKey = null
   }
 
-  setNetwork(network: Network): XChainClient {
+  setNetwork = (network: Network): XChainClient => {
     this.network = network
     this.thorClient = new CosmosSDKClient({
       server: this.getClientUrl(),
@@ -72,7 +72,7 @@ class Client implements ThorchainClient, XChainClient {
     return this
   }
 
-  getNetwork(): Network {
+  getNetwork = (): Network => {
     return this.network
   }
 
@@ -206,41 +206,56 @@ class Client implements ThorchainClient, XChainClient {
     }
   }
 
-  deposit = async ({ asset, amount, memo, fee }: DepositParam): Promise<TxHash> => {
+  private getDefaultStdTxFee = (): StdTxFee => {
+    return {
+      amount: [],
+      gas: '10000000',
+    }
+  }
+
+  deposit = async ({ asset, amount, memo }: DepositParam): Promise<TxHash> => {
     try {
+      this.registerCodecs()
+
+      if (!asset) asset = AssetRune
+
+      const assetBalance = await this.getBalance(this.getAddress(), asset)
+      const fee = await this.getFees()
+      if (assetBalance.length === 0 || assetBalance[0].amount.amount().lt(amount.amount().plus(fee.average.amount()))) {
+        throw new Error('insufficient funds')
+      }
+
+      const signer = this.getAddress()
+
       const msg: Msg = [
         MsgNativeTx.fromJSON({
           coins: [
             {
-              denom: getDenom(asset || AssetRune),
+              asset: getDenomWithChain(asset),
               amount: amount.amount().toString(),
             },
           ],
           memo,
-          signer: this.getAddress(),
+          signer,
         }),
       ]
 
-      console.log('msg', msg)
-
-      const signatures: StdTxSignature[] = []
-
       const unsignedStdTx = StdTx.fromJSON({
         msg,
-        fee: fee || {
-          gas: '200000',
-          amount: [],
-        },
-        signatures,
-        memo,
+        fee: this.getDefaultStdTxFee(),
+        signatures: [],
+        memo: '',
       })
 
-      const signer = AccAddress.fromBech32(this.getAddress())
+      const transferResult = await this.thorClient.signAndBroadcast(
+        unsignedStdTx,
+        this.getPrivateKey(),
+        AccAddress.fromBech32(signer),
+      )
 
-      console.log('unsignedStdTx', unsignedStdTx)
-      console.log('signer', signer)
-
-      const transferResult = await this.thorClient.signAndBroadcast(unsignedStdTx, this.getPrivateKey(), signer)
+      if (!isBroadcastSuccess(transferResult)) {
+        throw new Error(`failed to broadcast transaction: ${transferResult.txhash}`)
+      }
 
       return transferResult?.txhash || ''
     } catch (error) {
@@ -252,14 +267,27 @@ class Client implements ThorchainClient, XChainClient {
     try {
       this.registerCodecs()
 
+      if (!asset) asset = AssetRune
+
+      const assetBalance = await this.getBalance(this.getAddress(), asset)
+      const fee = await this.getFees()
+      if (assetBalance.length === 0 || assetBalance[0].amount.amount().lt(amount.amount().plus(fee.average.amount()))) {
+        throw new Error('insufficient funds')
+      }
+
       const transferResult = await this.thorClient.transfer({
         privkey: this.getPrivateKey(),
         from: this.getAddress(),
         to: recipient,
         amount: amount.amount().toString(),
-        asset: getDenom(asset || AssetRune),
+        asset: getDenom(asset),
         memo,
+        fee: this.getDefaultStdTxFee(),
       })
+
+      if (!isBroadcastSuccess(transferResult)) {
+        throw new Error(`failed to broadcast transaction: ${transferResult.txhash}`)
+      }
 
       return transferResult?.txhash || ''
     } catch (error) {
@@ -269,20 +297,16 @@ class Client implements ThorchainClient, XChainClient {
 
   // there is no fixed fee, we set fee amount when creating a transaction.
   getFees = async (): Promise<Fees> => {
-    return Promise.resolve({
-      type: 'base',
-      fast: baseAmount(750, DECIMAL),
-      fastest: baseAmount(2500, DECIMAL),
-      average: baseAmount(0, DECIMAL),
-    })
+    return Promise.resolve(this.getDefaultFees())
   }
 
   getDefaultFees = (): Fees => {
+    const fee = baseAmount(10000000, DECIMAL)
     return {
       type: 'base',
-      fast: baseAmount(750, DECIMAL),
-      fastest: baseAmount(2500, DECIMAL),
-      average: baseAmount(0, DECIMAL),
+      fast: fee,
+      fastest: fee,
+      average: fee,
     }
   }
 }